Add files via upload
Let's suppose that the forbidden cell does not affect the shortest path. In that case, the answer would be |𝑥𝐴−𝑥𝐵|+|𝑦𝐴−𝑦𝐵|.

The forbidden cell blocks the shortest path if and only if it belongs to every shortest path. In other words, if there is only one shortest path and the forbidden cell belongs to it. So, the answer differs from |𝑥𝐴−𝑥𝐵|+|𝑦𝐴−𝑦𝐵| if and only if 𝐴 and 𝐵 are in one row or column and 𝐹 is between them. In that case, the answer is |𝑥𝐴−𝑥𝐵|+|𝑦𝐴−𝑦𝐵|+2 (i.e. greater by 2).

In order to check that point 𝑅 lays between 𝑃 and 𝑄 on a straight line, just check min(𝑃,𝑄)<𝑅<max(𝑃,𝑄).